Processes .nitro files and renames based on content
Updates the file upload process to:

- Temporarily saves uploaded .nitro files.
- Extracts information from the .nitro file, including an internal name.
- Renames the uploaded file based on the internal name (if available) instead of the original filename.
- Includes error handling to remove temporary files if processing or renaming fails.

Also updates the NitroLayer Z value to be a float.

+++ backend/nitro.go
@@ -291,7 +291,7 @@ type NitroVisualization struct {
 }
 
 type NitroLayer struct {
-	Z           int     `json:"z"`
+	Z           float64 `json:"z"`
 	Alpha       int     `json:"alpha"`
 	Ink         string  `json:"ink"`
 	IgnoreMouse bool    `json:"ignoreMouse"`
